requirements.txt
pandas==2.3.3
uvicorn==0.38.0
numpy==2.2.6
fastapi==0.95.2
joblib==1.5.2
scikit-learn==1.7.0
sklearn-pandas==2.2.0
tensorflow==2.20.0
keras==3.12.0
scikeras==0.13.0
pydantic==1.10.13